loire-common: camera: Add support for the Camera Augmented Sensing Helper RGBC-IR
Support is added for CASH RGBC-IR. To enable it,
it is needed to declare TARGET_USES_CASH_EXTENSION.

RGBC-IR is only used when flash is disabled and AE
is enabled.

In such a case, AE will be disabled and the ISO and
exposure time values coming from CASH will be used
instead.

Change-Id: I0b643fe5159cbac3b7a296d4caf65bf518b6a814