loire-common: sepolicy: Don't audit hwservice add -> find.
Since an additional check related to getting the transport of a service
before registering it is more visible (this is moving a VTS test failure
to boot time for more visibility).

When adding a 'find' check on the 'add' path, this triggered some logs
here. 'find' has always been given with 'add' for other services, but
for a dontaudit, find has to be manually added.

Bug: 139274536
Test: TH
Change-Id: Ica4266dfaf17acc8c7bc8bd83054aa02811a4a37